htm: Seperate out the htm type
